scrap/requirements.txt
cssselect==0.9.2
Cython==0.24.1
lxml==3.6.4
peewee==2.8.3
Pillow==3.3.1
PyMySQL==0.7.9
pyquery==1.2.13
redis==2.10.5
requests==2.11.1
selenium==2.53.6
six==1.10.0
tesserocr==2.1.2
